[FIX] base_comment_template: Consider multi-company and partners
In a multi-company context, if you put a template with company_id set,
this template is trying to be assigned to any document, no matter which
company it belongs, as it runs in a sudo environment, or at least, you
can have several companies seen at the same time.

We add the company discriminant for avoiding this problem.

Furthermore, the previous code was adding a domain where the right
part may be `False` with the operator `in`, provoking an unexpected
behavior in the search.

This has been fixed stating directly the expected domain.

TT42591