Add possibility to use watermark on duplicated reports
Currently when duplicating a report, you can specify a different watermark for this duplicated report, but it will still use the watermark from the original report. This commit adds a feature, to use a watermark in a duplicated report, by adding "_duplicate" in the reports name. This is for example useful, if you want to print your sale report with different watermarks