[IMP] report_py3o: add new mixin model
Extending a model with this mixin allows to set a py3o template at model
and record level.

Using a common report it is possible to print a custom py3o template by
setting that template at model and record level.
A use case could be to want to print the invoice reports of a contact by
customising those reports using a py3o template. You would extend the
res.partner model to be able to add the template to that contact and
have a related on the invoices to print that custom report.

TT55652